requirements.txt
streamlit
langchain
langchain-community
langchain-huggingface
langchain-groq
faiss-cpu
python-docx
python-dotenv
huggingface-hub
sentence-transformers